advanced searching options - added spinners
