+++ app/src/main/java/com/alyjak/reposearch/events/MakeSearchEvent.kt
@@ -1,3 +1,10 @@
 package com.alyjak.reposearch.events
 
-class MakeSearchEvent(val query: String)
+import com.alyjak.reposearch.network.enums.Order
+import com.alyjak.reposearch.network.enums.SortingStrategy
+
+class MakeSearchEvent(
+    val query: String,
+    val sortingStrategy: SortingStrategy? = null,
+    val order: Order? = null
+)
