+++ app/src/main/java/com/alyjak/reposearch/network/GitHubRepository.kt
@@ -1,6 +1,8 @@
 package com.alyjak.reposearch.network
 
 import androidx.lifecycle.MutableLiveData
+import com.alyjak.reposearch.network.enums.Order
+import com.alyjak.reposearch.network.enums.SortingStrategy
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
@@ -8,18 +10,10 @@ class GitHubRepository {
 
     var results: MutableLiveData<Model.SearchingResult> = MutableLiveData()
 
-    suspend fun getSearchingResult() {
+    suspend fun getSearchingResult(query: String, sortingStrategy: SortingStrategy?, order: Order?) {
         withContext(Dispatchers.IO) {
             results.postValue(Network.gitHubService
-                .getRepositories("tetris")
-                .await())
-        }
-    }
-
-    suspend fun getSearchingResult(query: String) {
-        withContext(Dispatchers.IO) {
-            results.postValue(Network.gitHubService
-                .getRepositories(query)
+                .getRepositories(q = query, sort = sortingStrategy?.sortType, order = order?.order)
                 .await())
         }
     }
