+++ app/src/main/java/com/alyjak/reposearch/network/GitHubService.kt
@@ -1,5 +1,7 @@
 package com.alyjak.reposearch.network
 
+import com.alyjak.reposearch.network.enums.Order
+import com.alyjak.reposearch.network.enums.SortingStrategy
 import kotlinx.coroutines.Deferred
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -13,17 +15,5 @@ interface GitHubService {
         @Query("order") order: String? = Order.DESC.order
     ): Deferred<Model.SearchingResult>
 
-}
-
-enum class SortingStrategy(val sortType: String) {
-    STARS("stars"),
-    FORKS("forks"),
-    HELP_WANTED_ISSUES("help-wanted-issues"),
-    UPDATED("updated")
-}
-
-enum class Order(val order: String) {
-    DESC("desc"),
-    ASC("asc")
 }
