+++ app/src/main/java/com/alyjak/reposearch/network/enums/Order.kt
@@ -1,0 +1,6 @@
+package com.alyjak.reposearch.network.enums
+
+enum class Order(val order: String) {
+    DESC("desc"),
+    ASC("asc")
+}
