+++ app/src/main/java/com/alyjak/reposearch/network/enums/SortingStrategy.kt
@@ -1,0 +1,8 @@
+package com.alyjak.reposearch.network.enums
+
+enum class SortingStrategy(val sortType: String) {
+    STARS("stars"),
+    FORKS("forks"),
+    HELP_WANTED_ISSUES("help-wanted-issues"),
+    UPDATED("updated")
+}
